src/main/java/com/n26/controllers/TransactionController.java
package com.n26.controllers;

import com.n26.model.Transaction;
import com.n26.services.TransactionService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.*;

/**
 * Controller to handle transactions requests
 */


@RestController
@RequestMapping(value = "/transactions", produces = "application/json")
public class TransactionController {

	private TransactionService transactionService;

	@Autowired
	public TransactionController(TransactionService transactionService) {
		this.transactionService = transactionService;
	}

	@PostMapping
	@ResponseStatus(HttpStatus.CREATED)
	public void add(@RequestBody Transaction transaction) {
		this.transactionService.add(transaction);
	}

	@DeleteMapping
	@ResponseStatus(HttpStatus.NO_CONTENT)
	public void delete() {
		this.transactionService.delete();
	}

}
